Not Null,  Code refactoring
