+++ Yummi.Core/DTOs/RecipeDto.cs
@@ -3,10 +3,10 @@
     public class RecipeDto
     {
         public string Name { get; set; } = null!;
-        public string? Description { get; set; }
+        public string Description { get; set; } = null!;
         public string Instructions { get; set; } = null!;
-        public string? Cuisine { get; set; }
-        public string? Course { get; set; }
+        public string Cuisine { get; set; } = null!;
+        public string Course { get; set; } = null!;
         public int? PrepTime { get; set; }
         public int? CookTime { get; set; }
     }
